requirements.txt
amqp==2.5.2
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.2.3
astroid==2.3.3
awscli==1.18.97
bcrypt==3.1.7
beautifulsoup4==4.9.1
billiard==3.6.2.0
boto3==1.14.20
botocore==1.17.20
celery==4.4.0
certifi==2020.4.5.1
cffi==1.14.0
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
DateTime==4.3
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.1
Django==2.2
django-admin-rangefilter==0.5.2
django-celery-beat==2.1.0
django-celery-results==1.2.1
django-cors-headers==3.2.0
django-crispy-forms==1.6.0
django-extensions==2.2.9
django-filter==2.2.0
django-import-export==2.0
django-modeladmin-reorder==0.3.1
django-pagedown==0.1.1
django-rest-swagger==2.2.0
django-suit @ https://github.com/darklow/django-suit/tarball/v2
django-timezone-field==4.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docopt==0.6.2
docutils==0.15.2
docx2pdf==0.1.7
et-xmlfile==1.0.1
Faker==4.0.3
filelock==3.0.12
httpie==2.1.0
identify==1.4.21
idna==2.9
importlib-metadata==1.5.0
importlib-resources==3.0.0
isort==4.3.21
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
kombu==4.6.7
lazy-object-proxy==1.4.3
lxml==4.5.1
Markdown==3.2.1
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
nodeenv==1.4.0
num2words==0.5.10
numpy==1.19.2
oauthlib==3.1.0
odfpy==1.4.0
openapi-codec==1.3.2
opencv-python==4.4.0.44
openpyxl==3.0.2
packaging==20.4
pandas==1.1.3
Pillow==7.1.1
pluggy==0.13.1
pre-commit==2.6.0
psycopg2==2.8.5
py==1.9.0
pyasn1==0.4.8
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python-docx==0.8.10
python3-openid==3.1.0
pytz==2019.3
pywin32==228
PyYAML==5.2
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.5
s3transfer==0.3.3
simplejson==3.17.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.0
tablib==0.14.0
text-unidecode==1.3
toml==0.10.1
tox==3.16.1
tqdm==4.51.0
translate==3.5.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
virtualenv==20.0.27
virtualenvwrapper-win==1.2.6
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
zipp==3.0.0
zope.interface==5.1.0
